Closes #77
